Refactor popup
