+++ public/popup.css
@@ -202,3 +202,7 @@ h2 {
 .onoffswitch-checkbox:checked + .onoffswitch-label .onoffswitch-switch {
     right: 0px;
 }
+
+.latex-item#mathrm .symbol {
+    font-family: 'Times New Roman', Times, serif;
+}
