+++ src/utils/autoCommands.ts
@@ -1,0 +1,97 @@
+export interface AutoCommandsToSymbols {
+    [index: string]: string;
+}
+
+export const desmosDefualtAutoCommands: AutoCommandsToSymbols = {
+    sqrt: '√',
+    cbrt: '∛',
+    nthroot: '∜',
+    sum: 'Σ',
+    prod: '∏',
+    int: '∫',
+    percent: '%',
+    infinity: '∞',
+    infty: '∞',
+    alpha: 'α',
+    beta: 'β',
+    theta: 'θ',
+    pi: 'π',
+    tau: 'τ',
+    phi: 'ϕ',
+};
+
+export const basicAutoCommands: AutoCommandsToSymbols = {
+    gamma: 'γ',
+    digamma: 'ϝ',
+    Gamma: 'Γ',
+    delta: 'δ',
+    Delta: 'Δ',
+    epsilon: 'ϵ',
+    //varepsilon: 'ε',
+    zeta: 'ζ',
+    eta: 'η',
+    iota: 'ι',
+    kappa: 'κ',
+    //varkappa: 'ϰ',
+    lambda: 'λ',
+    Lambda: 'Λ',
+    mu: 'μ',
+    nu: 'ν',
+    xi: 'ξ',
+    //varpi: 'ϖ',
+    Pi: 'Π',
+    rho: 'ρ',
+    //varrho: 'ϱ',
+    sigma: 'σ',
+    //varsigma: 'ς',
+    Sigma: 'Σ',
+    upsilon: 'υ',
+    Upsilon: 'Υ',
+    //varphi: 'φ',
+    Phi: 'Φ',
+    chi: 'χ',
+    psi: 'ψ',
+    Psi: 'Ψ',
+    omega: 'ω',
+    Omega: 'Ω',
+    square: '□',
+    mid: '|',
+    parallel: '∥',
+    nparallel: '∦',
+    perp: '⊥',
+    pm: '±',
+    mp: '∓',
+    to: '→',
+    le: '≤',
+    ge: '≥',
+    ne: '≠',
+    times: '×',
+    div: '÷',
+    approx: '≈',
+};
+
+export const advancedAutoCommands = {
+    dot: ' ̇',
+    hat: 'ˆ',
+    vec: ' ⃗',
+    bar: '¯',
+    tidle: ' ̃',
+    langle: '⟨',
+    rangle: '⟩',
+    lVert: '||',
+    rVert: '||',
+    binom: '( )',
+    coprod: '∐',
+    frac: '½',
+    underline: '⎯',
+    overline: '¯',
+    overarc: '⌒',
+    mathrm: 'rm',
+    mathit: '𝑖𝑡',
+    mathbf: '𝐛𝐟',
+    mathsf: 'sf',
+    overrightarrow: '→',
+    overleftarrow: '←',
+    overleftrightarrow: '⟷',
+    MathQuillMathField: '▭',
+};
